Adds JSON encoded data format
The JSON representation of gron data makes
it smoother to grep the path component of the
entries, by virtue of not being rife with '[' and ']'
characters (unlike stadard gron data can be), which
are special in regex.

Also it makes it easy to create more complex filters,
as its easy for such programs to import and to emit
JSON represention gron of data.