Allow running as ungron to perform --ungron automatically
The executable is normally executed as `gron` which performs the gron
operation. However, in order to install it as `ungron`, an alias is not
necessary as a symbolic link can be used instead. When the program
detects that it is being run as an executable called `ungron` then it
will automatically set the `--ungron` argument.