Adds "decode all" option
Adds: `-a`, `-all` flag which means "decode all the objects,
pretending it's a JSON stream even if it's not actually."

Rationale: `gron` only decodes the first object, `gron -s`
requires a "correctly" formatted JSON stream (one object per
line), but it's not uncommon to get multiple objects per line
with tools that don't support JSON stream formatting.

This does require a positionable stream, however, since the
JSON decoder can read past the end of an object to be sure its
parsed correctly.  `io.Seekable` doesn't work, unfortunately,
because whilst we know where we want to be (`d.InputOffset()`),
we don't actually know where we currently are which precludes
the use of `io.SeekCurrent` and, bizarrely, it turns out that
`io.SeekSet` gets progressively slower as you seek further and
further into your (in this case) `bytes.Buffer`.

Thus we keep track of where we want to be (`moved`) and create
a `bytes.NewReader` for each attempted decode at the correct
position.  Crufty, definitely, and memory-allocation heavy,
probably, but it works and is surprisingly not that bad even
on large files.

My test 85MB JSON single line input takes ~64s (x86_64),
~43s (arm64) and ~275M to parse into 1024 objects comprising
1GB of output text.  Compare to `jq`: ~25s (x86_64),
~11s (arm64) using ~630M giving 350MB of output.